Week 2 commit - consumer and db

+++ producer/src/main/java/com/threatsense/producer/service/LogGeneratorService.java
@@ -1,7 +1,7 @@
 package com.threatsense.producer.service;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.threatsense.producer.model.NetworkLog;
+import com.threatsense.model.NetworkLog;
 import org.springframework.kafka.core.KafkaTemplate;
 import org.springframework.stereotype.Service;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
@@ -28,7 +28,7 @@ private final ObjectMapper mapper = new ObjectMapper().registerModule(new JavaTi
             log.destIP = (8 + rnd.nextInt(200)) + "." + rnd.nextInt(256) + "." + rnd.nextInt(256) + "." + rnd.nextInt(256);
             log.port = new int[]{22,80,443,3389}[rnd.nextInt(4)];
             log.protocol = rnd.nextDouble() < 0.85 ? "TCP" : "UDP";
-            log.bytes = 200 + rnd.nextInt(20000);
+            log.bytes = 200 + rnd.nextLong(20000);
             log.timestamp = Instant.now();
 
             String json = mapper.writeValueAsString(log);
